fix(unpack+pack): parsing and building with dynamic offsets ---
Move all parsing and building related functions (pack*, unpack*, sizeof) into a separate file for easy import handling.

Refs: #57
