+++ src/caterpillar/model/_struct.py
@@ -16,11 +16,8 @@
 import inspect
 import dataclasses as dc
 
-from tempfile import TemporaryFile
-from io import BytesIO, IOBase
-from collections import OrderedDict
+from io import BytesIO
 from collections.abc import Collection, Iterable
-from shutil import copyfileobj
 from typing import Any, Callable, Generic, Literal, ParamSpec, TypeVar
 from types import TracebackType
 from typing_extensions import (
@@ -32,15 +29,8 @@ from typing_extensions import (
     Buffer,
 )
 
-from caterpillar.byteorder import (
-    O_DEFAULT_ARCH,
-    O_DEFAULT_ENDIAN,
-    LittleEndian,
-    system_arch,
-)
-from caterpillar.shared import ATTR_PACK, getstruct, hasstruct, ATTR_STRUCT
-from caterpillar.context import O_CONTEXT_FACTORY, CTX_STREAM, Context
-from caterpillar.exception import DynamicSizeError, InvalidValueError
+from caterpillar.shared import getstruct, hasstruct, ATTR_STRUCT
+from caterpillar.exception import InvalidValueError
 from caterpillar.options import (
     S_EVAL_ANNOTATIONS,
     S_UNION,
@@ -51,22 +41,16 @@ from caterpillar.options import (
 )
 from caterpillar.fields import Field, INVALID_DEFAULT
 from caterpillar import registry
-from caterpillar.shared import MODE_PACK, MODE_UNPACK
 from caterpillar.abc import (
-    _ContainsStruct,
-    _OT,
-    _IT,
-    _SupportsPack,
     _StreamType,
-    _SupportsUnpack,
     _ContextLike,
     _StructLike,
-    _SupportsSize,
     _OptionLike,
     _EndianLike,
     _ArchLike,
     _LengthT,
 )
+from .provider import unpack, pack, unpack_file, pack_into, sizeof
 from ._base import Sequence
 
 
@@ -606,6 +590,7 @@ class struct_factory:
 
 struct = struct_factory.new
 
+
 class UnionHook(Generic[_ModelT]):
     """Implementation of a hook to simulate union types.
 
@@ -771,565 +756,3 @@ def union(
         )
 
     return wrap
-
-
-@overload
-def pack(
-    obj: _ContainsStruct[_IT, _OT],
-    struct: None = None,
-    /,
-    *,
-    use_tempfile: bool = False,
-    as_field: bool = False,
-    order: _EndianLike | None = None,
-    arch: _ArchLike | None = None,
-    **kwargs: Any,
-) -> bytes: ...
-@overload
-def pack(
-    # fallback to allow arbitrary objects
-    obj: object,
-    struct: None = None,
-    /,
-    *,
-    use_tempfile: bool = False,
-    as_field: bool = False,
-    order: _EndianLike | None = None,
-    arch: _ArchLike | None = None,
-    **kwds: Any,
-) -> bytes: ...
-@overload
-def pack(
-    obj: _IT,
-    struct: _SupportsPack[_IT],
-    /,
-    *,
-    use_tempfile: bool = False,
-    as_field: bool = False,
-    order: _EndianLike | None = None,
-    arch: _ArchLike | None = None,
-    **kwargs: Any,
-) -> bytes: ...
-@overload
-def pack(
-    obj: _IT,
-    # support arbitrary struct types
-    struct: type[_IT],
-    /,
-    *,
-    use_tempfile: bool = False,
-    as_field: bool = False,
-    order: _EndianLike | None = None,
-    arch: _ArchLike | None = None,
-    **kwargs: Any,
-) -> bytes: ...
-
-
-# @overload
-# def pack(
-#     obj: _IT,
-#     struct: _ContainsStruct[_IT, _OT],
-#     /,
-#     **kwargs: Any,
-# ) -> bytes: ...
-def pack(
-    obj: _IT | _ContainsStruct[_IT, _OT],
-    struct: _SupportsPack[_IT] | type[_IT] | _ContainsStruct[_IT, _OT] | None = None,
-    /,
-    *,
-    use_tempfile: bool = False,
-    as_field: bool = False,
-    order: _EndianLike | None = None,
-    arch: _ArchLike | None = None,
-    **kwargs: Any,
-) -> bytes:
-    """
-    Pack an object into a bytes buffer using the specified struct.
-
-    :param obj: The object to pack.
-    :param struct: The struct to use for packing.
-    :param kwds: Additional keyword arguments to pass to the pack function.
-
-    :return: The packed bytes.
-    """
-    buffer = BytesIO()
-    pack_into(  # pyright: ignore[reportCallIssue]
-        obj,
-        buffer,
-        struct,  # pyright: ignore[reportArgumentType]
-        order=order,
-        arch=arch,
-        use_tempfile=use_tempfile,
-        as_field=as_field,
-        **kwargs,
-    )
-    return buffer.getvalue()
-
-
-@overload
-def pack_into(
-    obj: _ContainsStruct[_IT, _OT],
-    buffer: _StreamType,
-    struct: None = None,
-    /,
-    *,
-    use_tempfile: bool = ...,
-    as_field: bool = ...,
-    order: _EndianLike | None = ...,
-    arch: _ArchLike | None = None,
-    **kwds: Any,
-) -> None: ...
-@overload
-def pack_into(
-    obj: object,
-    buffer: _StreamType,
-    struct: None = None,
-    /,
-    *,
-    use_tempfile: bool = ...,
-    as_field: bool = ...,
-    order: _EndianLike | None = ...,
-    arch: _ArchLike | None = None,
-    **kwds: Any,
-) -> None: ...
-@overload
-def pack_into(
-    obj: _IT,
-    buffer: _StreamType,
-    struct: _SupportsPack[_IT],
-    /,
-    *,
-    use_tempfile: bool = ...,
-    as_field: bool = ...,
-    order: _EndianLike | None = ...,
-    arch: _ArchLike | None = None,
-    **kwds: Any,
-) -> None: ...
-@overload
-def pack_into(
-    obj: _IT,
-    buffer: _StreamType,
-    struct: type[_IT],
-    /,
-    *,
-    use_tempfile: bool = ...,
-    as_field: bool = ...,
-    order: _EndianLike | None = ...,
-    arch: _ArchLike | None = None,
-    **kwds: Any,
-) -> None: ...
-@overload
-def pack_into(
-    obj: _IT,
-    buffer: _StreamType,
-    struct: _ContainsStruct[_IT, _OT],
-    /,
-    *,
-    use_tempfile: bool = ...,
-    as_field: bool = ...,
-    order: _EndianLike | None = ...,
-    arch: _ArchLike | None = None,
-    **kwds: Any,
-) -> None: ...
-def pack_into(
-    obj: _IT | _ContainsStruct[_IT, _OT],
-    buffer: _StreamType,
-    struct: type[_IT] | _ContainsStruct[_IT, _OT] | _SupportsPack[_IT] | None = None,
-    /,
-    *,
-    use_tempfile: bool = False,
-    as_field: bool = False,
-    order: _EndianLike | None = None,
-    arch: _ArchLike | None = None,
-    **kwds: Any,
-) -> None:
-    """
-    Pack an object into the specified buffer using the specified struct.
-
-    This function serializes an object (`obj`) into a given buffer, using
-    a struct to define how the object should be packed. Optionally, the
-    function can handle temporary files for packing, use a `Field` wrapper
-    around the struct, and support additional keyword arguments. The packed
-    data is written to the `buffer`.
-
-    Example 1: Packing an object into a bytes buffer
-
-    >>> buffer = BytesIO()
-    >>> my_obj = SomeObject()  # Assume SomeObject is a valid object to be packed
-    >>> pack_into(my_obj, buffer, struct=SomeStruct())  # Using a specific struct
-    >>> buffer.getvalue()
-    b"..."
-
-    Example 2: Packing into a file-like stream (e.g., file)
-
-    >>> with open('packed_data.bin', 'wb') as f:
-    ...     pack_into(my_obj, f, struct=SomeStruct())  # Pack into a file
-
-    Example 3: Using `as_field` to wrap the struct in a Field before packing
-
-    >>> buffer = BytesIO()
-    >>> pack_into(42, buffer, struct=uint8, as_field=True)
-    >>> buffer.getvalue()
-    b"\\x2a"
-
-    :param obj: The object to pack (could be a plain object or a structure-like object).
-    :param buffer: The buffer to pack the object into (a writable stream such as `BytesIO` or a file).
-    :param struct: The struct to use for packing. If not specified, will infer from `obj`.
-    :param use_tempfile: Whether to use a temporary file for packing (experimental).
-    :param as_field: Whether to wrap the struct in a `Field` before packing.
-    :param kwds: Additional keyword arguments to pass to the pack function.
-
-    :raises TypeError: If no `struct` is specified and cannot be inferred from the object.
-    """
-    # fmt: off
-    offsets: OrderedDict[int, int] = OrderedDict()
-    # NOTE: we don't have to set _root here because the default root context
-    # will be this instance.
-    context = (O_CONTEXT_FACTORY.value or Context)(
-        _parent=None,
-        _path="<root>",
-        _pos=0,
-        _offsets=offsets,
-        _is_seq=False,
-        _order=order or O_DEFAULT_ENDIAN.value or LittleEndian,
-        _arch=arch or O_DEFAULT_ARCH.value or system_arch,
-        mode=MODE_PACK,
-        **kwds,
-    )
-    if struct is None:
-        struct = getstruct(obj)
-    elif as_field:
-        struct = Field(struct)  # pyright: ignore[reportArgumentType]
-    elif hasstruct(struct):
-        struct = getstruct(struct)
-
-    if struct is None:
-        raise TypeError("struct must be specified")
-
-    if not hasattr(struct, ATTR_PACK):
-        raise TypeError(
-            f"pack* called with an unknown struct type ({type(struct)}) - "
-            + "no __pack__ defined!"
-        )
-
-    prev_order = O_DEFAULT_ENDIAN.value
-    prev_arch = O_DEFAULT_ARCH.value
-    if order:
-        O_DEFAULT_ENDIAN.value = order
-    if arch:
-        O_DEFAULT_ARCH.value = prev_arch
-    try:
-        start = 0
-        if use_tempfile:
-            # NOTE: this implementation is exprimental - use this option with caution.
-            with TemporaryFile() as stream:
-                context[CTX_STREAM] = stream
-                struct.__pack__(obj, context)  # pyright: ignore[reportAttributeAccessIssue, reportArgumentType]
-
-                for offset, value in offsets.items():
-                    stream.seek(start)  # pyright: ignore[reportUnusedCallResult]
-                    buffer.write(stream.read(offset - start))
-                    buffer.write(value)
-                    start = offset
-                if len(offsets) == 0:
-                    stream.seek(0)  # pyright: ignore[reportUnusedCallResult]
-                    copyfileobj(stream, buffer)
-
-        else:
-            # Default implementation: We use an in-memory buffer to store all packed
-            # elements and then apply all offset-packed objects.
-            stream = BytesIO()
-            context[CTX_STREAM] = stream
-            struct.__pack__(obj, context) # pyright: ignore[reportAttributeAccessIssue, reportArgumentType]
-
-            content = stream.getbuffer()
-            if len(offsets) == 0:
-                buffer.write(content)
-            else:
-                for offset, value in offsets.items():
-                    buffer.write(content[start:offset])
-                    buffer.write(value)
-                    start = offset
-    finally:
-        O_DEFAULT_ENDIAN.value = prev_order
-        O_DEFAULT_ARCH.value = prev_arch
-
-
-@overload
-def pack_file(
-    obj: _ContainsStruct[_IT, _OT],
-    filename: str,
-    struct: None = None,
-    /,
-    *,
-    use_tempfile: bool = ...,
-    as_field: bool = ...,
-    order: _EndianLike | None = ...,
-    arch: _ArchLike | None = ...,
-    **kwds: Any,
-) -> None: ...
-@overload
-def pack_file(
-    obj: object,
-    filename: str,
-    struct: None = None,
-    /,
-    *,
-    use_tempfile: bool = ...,
-    as_field: bool = ...,
-    order: _EndianLike | None = ...,
-    arch: _ArchLike | None = ...,
-    **kwds: Any,
-) -> None: ...
-@overload
-def pack_file(
-    obj: _IT,
-    filename: str,
-    struct: _SupportsPack[_IT],
-    /,
-    *,
-    use_tempfile: bool = ...,
-    as_field: bool = ...,
-    order: _EndianLike | None = ...,
-    arch: _ArchLike | None = ...,
-    **kwds: Any,
-) -> None: ...
-@overload
-def pack_file(
-    obj: _IT,
-    filename: str,
-    struct: type[_IT],
-    /,
-    *,
-    use_tempfile: bool = ...,
-    as_field: bool = ...,
-    order: _EndianLike | None = ...,
-    arch: _ArchLike | None = ...,
-    **kwds: Any,
-) -> None: ...
-@overload
-def pack_file(
-    obj: _IT,
-    filename: str,
-    struct: _ContainsStruct[_IT, _OT],
-    /,
-    *,
-    use_tempfile: bool = ...,
-    as_field: bool = ...,
-    order: _EndianLike | None = ...,
-    arch: _ArchLike | None = ...,
-    **kwds: Any,
-) -> None: ...
-def pack_file(
-    obj: _IT | _ContainsStruct[_IT, _OT],
-    filename: str,
-    struct: _SupportsPack[_IT] | type[_IT] | _ContainsStruct[_IT, _OT] | None = None,
-    /,
-    *,
-    use_tempfile: bool = False,
-    as_field: bool = False,
-    order: _EndianLike | None = None,
-    arch: _ArchLike | None = None,
-    **kwds: Any,
-) -> None:
-    """
-    Pack an object into a file using the specified struct.
-
-    :param obj: The object to pack.
-    :param filename: The name of the file to write to.
-    :param struct: The struct to use for packing.
-    :param kwds: Additional keyword arguments to pass to the pack function.
-
-    :return: None
-    """
-    with open(filename, "w+b") as fp:
-        pack_into(  # pyright: ignore[reportCallIssue]
-            obj,
-            fp,
-            struct,  # pyright: ignore[reportArgumentType]
-            use_tempfile=use_tempfile,
-            as_field=as_field,
-            order=order,
-            arch=arch,
-            **kwds,
-        )
-
-
-@overload
-def unpack(
-    struct: _ContainsStruct[_IT, _OT],
-    buffer: Buffer | _StreamType,
-    /,
-    *,
-    as_field: bool = False,
-    order: _EndianLike | None = None,
-    arch: _ArchLike | None = None,
-    **kwds: Any,
-) -> _OT: ...
-@overload
-def unpack(
-    struct: _SupportsUnpack[_OT],
-    buffer: Buffer | _StreamType,
-    /,
-    *,
-    as_field: bool = False,
-    order: _EndianLike | None = None,
-    arch: _ArchLike | None = None,
-    **kwds: Any,
-) -> _OT: ...
-@overload
-def unpack(
-    struct: type[_OT],
-    buffer: Buffer | _StreamType,
-    /,
-    *,
-    as_field: bool = False,
-    order: _EndianLike | None = None,
-    arch: _ArchLike | None = None,
-    **kwds: Any,
-) -> _OT: ...
-def unpack(
-    struct: type[_OT] | _SupportsUnpack[_OT] | _ContainsStruct[_IT, _OT],
-    buffer: Buffer | _StreamType,
-    /,
-    as_field: bool = False,
-    order: _EndianLike | None = None,
-    arch: _ArchLike | None = None,
-    **kwds: Any,
-) -> _OT:
-    """
-    Unpack an object from a bytes buffer or stream using the specified struct.
-
-    This function takes a `struct` that defines how data should be unpacked,
-    a `buffer` (either bytes or a stream) containing the serialized data, and
-    returns the unpacked object. If `as_field` is set to True, the `struct` is
-    wrapped by a `Field`. Additional keyword arguments are passed to the root
-    context as attributes.
-
-    Example:
-
-    >>> buffer = b'\\x00\\x01\\x02\\x03'
-    >>> struct = SomeStruct()
-    >>> unpack(struct, buffer)
-    ...
-
-    :param struct: The struct to use for unpacking (could be a `SupportsUnpack` or `ContainsStruct` object).
-    :param buffer: The bytes buffer or stream to unpack from.
-    :param as_field: Whether to wrap the struct in a `Field` transformer before unpacking.
-    :param kwds: Additional keyword arguments to pass to the unpack function.
-
-    :return: The unpacked object, which is the result of calling `struct.__unpack__(context)`.
-
-    :raises TypeError: If the `struct` is not a valid struct instance.
-    """
-    # fmt: off
-    # prepare the data stream
-    stream = buffer if isinstance(buffer, IOBase) else BytesIO(buffer)
-    context = (O_CONTEXT_FACTORY.value or Context)(
-        _path="<root>",
-        _parent=None,
-        _io=stream,
-        _pos=0,
-        _is_seq=False,
-        _order=order or O_DEFAULT_ENDIAN.value or LittleEndian,
-        _arch=arch or O_DEFAULT_ARCH.value or system_arch,
-        mode=MODE_UNPACK,
-        **kwds,
-    )
-    if as_field:
-        struct = Field(struct)  # pyright: ignore[reportArgumentType]
-    elif hasstruct(struct):
-        struct = getstruct(struct)
-
-    if not isinstance(struct, _SupportsUnpack):
-        raise TypeError(f"{type(struct).__name__} is not a valid struct instance!")
-
-    prev_order = O_DEFAULT_ENDIAN.value
-    prev_arch = O_DEFAULT_ARCH.value
-    if order:
-        O_DEFAULT_ENDIAN.value = order
-    if arch:
-        O_DEFAULT_ARCH.value = arch
-
-    try:
-        return struct.__unpack__(context)
-    finally:
-        O_DEFAULT_ARCH.value = prev_arch
-        O_DEFAULT_ENDIAN.value = prev_order
-
-
-@overload
-def unpack_file(
-    struct: _ContainsStruct[_IT, _OT],
-    filename: str,
-    /,
-    *,
-    as_field: bool = False,
-    order: _EndianLike | None = None,
-    arch: _ArchLike | None = None,
-    **kwds: Any,
-) -> _OT: ...
-@overload
-def unpack_file(
-    struct: _SupportsUnpack[_OT],
-    filename: str,
-    /,
-    *,
-    as_field: bool = False,
-    order: _EndianLike | None = None,
-    arch: _ArchLike | None = None,
-    **kwds: Any,
-) -> _OT: ...
-@overload
-def unpack_file(
-    struct: type[_OT],
-    filename: str,
-    /,
-    *,
-    as_field: bool = False,
-    order: _EndianLike | None = None,
-    arch: _ArchLike | None = None,
-    **kwds: Any,
-) -> _OT: ...
-def unpack_file(
-    struct: type[_OT] | _SupportsUnpack[_OT] | _ContainsStruct[_IT, _OT],
-    filename: str,
-    /,
-    *,
-    as_field: bool = False,
-    order: _EndianLike | None = None,
-    arch: _ArchLike | None = None,
-    **kwds: Any,
-):
-    """
-    Unpack an object from a file using the specified struct.
-
-    :param struct: The struct to use for unpacking.
-    :param filename: The name of the file to read from.
-    :param kwds: Additional keyword arguments to pass to the unpack function.
-
-    :return: The unpacked object.
-    """
-    with open(filename, "rb") as fp:
-        return unpack(struct, fp, as_field=as_field, arch=arch, order=order, **kwds)
-
-
-@overload
-def sizeof(obj: _SupportsSize, **kwds: Any) -> int: ...
-@overload
-def sizeof(obj: _ContainsStruct, **kwds: Any) -> int: ...
-@overload
-def sizeof(obj: type, **kwds: Any) -> int: ...
-def sizeof(obj: _SupportsSize | _ContainsStruct | type, **kwds: Any) -> int:
-    context = (O_CONTEXT_FACTORY.value or Context)(_parent=None, _path="<root>", **kwds)
-    struct_ = obj
-    if hasstruct(struct_):
-        struct_ = getstruct(struct_)
-
-    if not isinstance(struct_, _SupportsSize):
-        raise TypeError(f"{type(struct_).__name__} does not support size calculation!")
-
-    size = struct_.__size__(context)
-    if not isinstance(size, int):  # pyright: ignore[reportUnnecessaryIsInstance]
-        raise DynamicSizeError(f"Struct {struct_} is using a dynamic size ", context)
-    return size
